All checkers are loaded automatically from rebasehelper/checkers folder
